+++ packages/core/src/codewhisperer/models/model.ts
@@ -34,6 +34,10 @@ interface VsCodeState {
      */
     isCodeWhispererEditing: boolean
     /**
+     * Keeps track of whether or not recommendations are currently running
+     */
+    isRecommendationsActive: boolean
+    /**
      * Timestamp of previous user edit
      */
     lastUserModificationTime: number
@@ -44,6 +48,9 @@ interface VsCodeState {
 export const vsCodeState: VsCodeState = {
     isIntelliSenseActive: false,
     isCodeWhispererEditing: false,
+    // hack to globally keep track of whether or not recommendations are currently running. This allows us to know
+    // when recommendations have ran during e2e tests
+    isRecommendationsActive: false,
     lastUserModificationTime: 0,
     isFreeTierLimitReached: false,
 }
@@ -668,16 +675,15 @@ export enum BuildSystem {
     Unknown = 'Unknown',
 }
 
-// TO-DO: include the custom YAML file path here somewhere?
 export class ZipManifest {
     sourcesRoot: string = 'sources/'
     dependenciesRoot: string = 'dependencies/'
-    buildLogs: string = 'build-logs.txt'
     version: string = '1.0'
     hilCapabilities: string[] = ['HIL_1pDependency_VersionUpgrade']
-    // TO-DO: add 'CLIENT_SIDE_BUILD' here when releasing
-    // TO-DO: add something like AGENTIC_PLAN_V1 here when BE allowlists everyone
-    transformCapabilities: string[] = ['EXPLAINABILITY_V1']
+    transformCapabilities: string[] = ['EXPLAINABILITY_V1', 'SELECTIVE_TRANSFORMATION_V2', 'CLIENT_SIDE_BUILD']
+    noInteractiveMode: boolean = true
+    dependencyUpgradeConfigFile?: string = undefined
+    compilationsJsonFile: string = 'compilations.json'
     customBuildCommand: string = 'clean test'
     requestedConversions?: {
         sqlConversion?: {
@@ -775,7 +781,7 @@ export class TransformByQState {
 
     private polledJobStatus: string = ''
 
-    private jobFailureMetadata: string = ''
+    private hasSeenTransforming: boolean = false
 
     private payloadFilePath: string = ''
 
@@ -822,6 +828,10 @@ export class TransformByQState {
 
     public isPartiallySucceeded() {
         return this.transformByQState === TransformByQStatus.PartiallySucceeded
+    }
+
+    public getHasSeenTransforming() {
+        return this.hasSeenTransforming
     }
 
     public getTransformationType() {
@@ -916,10 +926,6 @@ export class TransformByQState {
         return this.projectCopyFilePath
     }
 
-    public getJobFailureMetadata() {
-        return this.jobFailureMetadata
-    }
-
     public getPayloadFilePath() {
         return this.payloadFilePath
     }
@@ -998,6 +1004,10 @@ export class TransformByQState {
 
     public setToPartiallySucceeded() {
         this.transformByQState = TransformByQStatus.PartiallySucceeded
+    }
+
+    public setHasSeenTransforming(hasSeen: boolean) {
+        this.hasSeenTransforming = hasSeen
     }
 
     public setTransformationType(type: TransformationType) {
@@ -1084,10 +1094,6 @@ export class TransformByQState {
         this.projectCopyFilePath = filePath
     }
 
-    public setJobFailureMetadata(data: string) {
-        this.jobFailureMetadata = data
-    }
-
     public setPayloadFilePath(payloadFilePath: string) {
         this.payloadFilePath = payloadFilePath
     }
@@ -1146,9 +1152,9 @@ export class TransformByQState {
 
     public setJobDefaults() {
         this.setToNotStarted()
+        this.hasSeenTransforming = false
         this.jobFailureErrorNotification = undefined
         this.jobFailureErrorChatMessage = undefined
-        this.jobFailureMetadata = ''
         this.payloadFilePath = ''
         this.metadataPathSQL = ''
         this.customVersionPath = ''