+++ packages/toolkit/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 3.67.0 2025-06-25
+
+- **Bug Fix** State Machine deployments can now be initiated directly from Workflow Studio without closing the editor
+- **Bug Fix** Step Function performance metrics now accurately reflect only Workflow Studio document activity
+- **Feature** AccessAnalyzer: CheckNoPublicAccess custom policy check supports additional resource types.
+
+## 3.66.0 2025-06-18
+
+- Miscellaneous non-user-facing changes
+
+## 3.65.0 2025-06-13
+
+- Miscellaneous non-user-facing changes
+
+## 3.64.0 2025-06-04
+
+- Miscellaneous non-user-facing changes
+
 ## 3.63.0 2025-05-22
 
 - Miscellaneous non-user-facing changes